backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlmodel>=0.0.21
sqlalchemy>=2.0
pydantic-settings>=2.4
python-dotenv>=1.0
psycopg2-binary
pgvector
httpx
python-multipart
requests
reportlab>=4.0.0
